+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
  - Add 12-volt DC fan control circuit to manual (@Theoi-Meteoroi) ([#184](https://github.com/kizniche/mycodo/issues/184)) (See [Schematics for DC Fan Control](https://github.com/kizniche/Mycodo/blob/master/mycodo-manual.md#schematics-for-dc-fan-control))
 
+### Bugfixes
+
+ - Fix PWM Signal Input ([#324](https://github.com/kizniche/mycodo/issues/324))
+
 
 ## 5.4.3 (2017-11-18)